hypr: fix stuff

+++ setup/hyprland/config/keyboard.sh
@@ -1,5 +1,7 @@
 # set qwerty layout for login prompt
-localectl set-x11-keymap us pc105
-mkdir -p $HOME/.xkb/symbols
-rm $HOME/.xkb/symbols/fr
-ln -s $DOTFILES_PATH/config/qwerty-l.xkb_symbols $HOME/.xkb/symbols/fr
+if [ $STATE_FLAG = "first-install" ]; then
+    localectl set-x11-keymap us pc105
+    mkdir -p $HOME/.xkb/symbols
+    rm $HOME/.xkb/symbols/fr
+    ln -s $DOTFILES_PATH/config/qwerty-l.xkb_symbols $HOME/.xkb/symbols/fr
+fi
